Correct bug in GitHub Oauth flow
