+++ pyproject.toml
@@ -96,4 +96,7 @@ dev = [
     "docker>=7.1.0",
     "pytest==8.3.3",
     "pytest-cov>=6.1.1",
+    "ruff>=0.12.0",
 ]
+
+[tool.ruff.lint]
