+++ tests/integration/oauth/constants.py
@@ -1,0 +1,1 @@
+GITHUB_OATH_LOGIC_PATCH_ROOT = "middleware.primary_resource_logic.github_oauth"
